Make proportional geometry f64 via Finite
ProportionalCoord was an f32 with its own hand-written Eq/Hash/Ord (to_bits /
total_cmp) and finiteness/-0.0 handling — the same shape GeoPoint and Meters
carried before Finite. Wrap it in Finite (f64) so those derive, and widen the
proportional-geometry grammar (ProportionalRect, ProportionalPoint,
ProportionalPolyline, ImageGeometry::bbox, SubimageRegion::rect) from f32 to f64
to match. The f32 was incidental — these are [0,1] normalized image coordinates
where f64 is ample — and one float type across the value grammar reads cleaner.

new() validates finiteness through Finite and keeps the [0,1] range check; the
custom Deserialize routes the number through new(). Three wire goldens
regenerate: the f32-imprecise decimals (0.30000001192092896) become clean f64
(0.3). Scope is core-only — the vision pipeline's bounding boxes are a separate
u32-pixel type, untouched.

Third of the Finite series (GeoPoint, Meters, ProportionalCoord); the scattered
validated-finite-float logic now lives in one place.

+++ core/src/grammar/composites.rs
@@ -116,7 +116,7 @@ pub enum SubimageRegion {
 impl SubimageRegion {
     /// Construct a rectangular subimage region. See
     /// [`ProportionalRect::new`] for the validation rules.
-    pub fn rect(ax: f32, ay: f32, bx: f32, by: f32) -> Result<Self, ProportionalCoordError> {
+    pub fn rect(ax: f64, ay: f64, bx: f64, by: f64) -> Result<Self, ProportionalCoordError> {
         Ok(Self::Rect {
             rect: ProportionalRect::new(ax, ay, bx, by)?,
         })
